quantum_ml_time_series_project/src/optimization.py

import numpy as np
from scipy.optimize import minimize
from qiskit import QuantumCircuit, Aer, transpile, execute

class QuantumOptimization:
    def __init__(self, num_qubits):
        self.num_qubits = num_qubits
        self.qc = QuantumCircuit(num_qubits)
        self.backend = Aer.get_backend('qasm_simulator')

    def optimize(self, objective_function, initial_params, bounds=None, method='COBYLA'):
        """Quantum-aware optimization method using classical optimization algorithms."""
        result = minimize(objective_function, initial_params, method=method, bounds=bounds)
        return result

    def quantum_variational_circuit(self, params):
        """Create a variational quantum circuit for optimization."""
        self.qc.rx(params[0], 0)
        self.qc.ry(params[1], 1)
        self.qc.cx(0, 1)
        transpiled_qc = transpile(self.qc, self.backend)
        return transpiled_qc

    def evaluate_quantum_circuit(self, quantum_circuit):
        """Evaluate the quantum circuit and return measurement results."""
        job = execute(quantum_circuit, backend=self.backend, shots=1024)
        result = job.result()
        counts = result.get_counts()
        return counts

    def optimize_variational_circuit(self, objective_function, initial_params, iterations=100):
        """Optimize a variational quantum circuit by iterating over a classical optimization method."""
        best_params = initial_params
        best_value = float('inf')

        for _ in range(iterations):
            result = self.optimize(objective_function, best_params)
            if result.fun < best_value:
                best_value = result.fun
                best_params = result.x

        return best_params

    def apply_penalty_term(self, params, penalty_factor=0.01):
        """Applies a penalty term to regularize the optimization process."""
        penalty = penalty_factor * np.sum(np.abs(params))
        return penalty

    def optimize_with_penalty(self, objective_function, initial_params, penalty_factor=0.01):
        """Optimize the function while applying a penalty to the parameters."""
        def penalized_function(params):
            return objective_function(params) + self.apply_penalty_term(params, penalty_factor)
        
        return self.optimize(penalized_function, initial_params)
